Move application listener from security component to app component
