+++ components/application-mgt/org.wso2.carbon.identity.application.mgt/pom.xml
@@ -89,6 +89,10 @@
             <groupId>org.wso2.carbon.consent.mgt</groupId>
             <artifactId>org.wso2.carbon.consent.mgt.core</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.wso2.carbon.identity.framework</groupId>
+            <artifactId>org.wso2.carbon.security.mgt</artifactId>
+        </dependency>
 
         <!-- Adding dependency to identity.core temporary -->
         <!-- 
@@ -153,7 +157,10 @@
                             org.wso2.carbon.context; version="${carbon.kernel.package.import.version.range}",
                             org.wso2.carbon.core; version="${carbon.kernel.package.import.version.range}",
 
-                            org.wso2.carbon.security.*; version="${carbon.identity.package.import.version.range}",
+                            org.wso2.carbon.security; version="${carbon.identity.package.import.version.range}",
+                            org.wso2.carbon.security.config; version="${carbon.identity.package.import.version.range}",
+                            org.wso2.carbon.security.sts.service; version="${carbon.identity.package.import.version.range}",
+                            org.wso2.carbon.security.sts.service.util; version="${carbon.identity.package.import.version.range}",
                             org.wso2.carbon.identity.base.*; version="${carbon.identity.package.import.version.range}",
                             org.wso2.carbon.identity.core.*; version="${carbon.identity.package.import.version.range}",
                             org.wso2.carbon.directory.server.manager; version="${carbon.identity.package.import.version.range}",
