+++ components/security-mgt/org.wso2.carbon.security.mgt/pom.xml
@@ -81,14 +81,6 @@
             <groupId>org.wso2.carbon.identity.framework</groupId>
             <artifactId>org.wso2.carbon.identity.core</artifactId>
         </dependency>
-        <dependency>
-            <groupId>org.wso2.carbon.identity.framework</groupId>
-            <artifactId>org.wso2.carbon.identity.application.mgt</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.wso2.carbon.identity.framework</groupId>
-            <artifactId>org.wso2.carbon.identity.application.common</artifactId>
-        </dependency>
     </dependencies>
 
     <build>
@@ -147,14 +139,7 @@
                             org.wso2.carbon.registry.core.*;version="${carbon.kernel.registry.imp.pkg.version}",
                             org.wso2.carbon.registry.api;version="${carbon.kernel.registry.imp.pkg.version}",
                             org.wso2.carbon.identity.core.*; version="${carbon.identity.package.import.version.range}",
-                            org.wso2.carbon.identity.base; version="${carbon.identity.package.import.version.range}",
-                            org.wso2.carbon.identity.core.*; version="${carbon.identity.package.import.version.range}",
-                            org.wso2.carbon.identity.application.common; version="${carbon.identity.package.import.version.range}",
-                            org.wso2.carbon.identity.application.common.model; version="${carbon.identity.package.import.version.range}",
-                            org.wso2.carbon.identity.application.common.util; version="${carbon.identity.package.import.version.range}",
-                            org.wso2.carbon.identity.application.mgt; version="${carbon.identity.package.import.version.range}",
-                            org.wso2.carbon.identity.application.mgt.listener; version="${carbon.identity.package.import.version.range}",
-                            org.wso2.carbon.identity.application.mgt.dao; version="${carbon.identity.package.import.version.range}"
+                            org.wso2.carbon.identity.base; version="${carbon.identity.package.import.version.range}"
                         </Import-Package>
                         <Export-Package>
                             !org.wso2.carbon.security.internal,
